feat: make physics shape update generic, increased spawn vel
